Make BaselineJavaVersions test to separate the two
